--- a/go.mod
+++ b/go.mod
@@ -7,3 +7,5 @@ require (
 	github.com/iancoleman/strcase v0.1.3
 	golang.org/x/exp v0.0.0-20220314205449-43aec2f8a4e7
 )
+
+require github.com/klauspost/compress v1.15.8 // indirect