--- a/net/packet/packet.go
+++ b/net/packet/packet.go
@@ -2,12 +2,32 @@ package packet
 
 import (
 	"bytes"
-	// "compress/zlib"
+	// "compress/zlib" // should use BestSpeed(level 1)
 	"fmt"
 	"io"
 
 	"github.com/klauspost/compress/zlib"
 )
+
+type ZlibWriter zlib.Writer
+
+func (zw *ZlibWriter) Reset(w io.Writer) {
+	(*zlib.Writer)(zw).Reset(w)
+}
+func (zw *ZlibWriter) Write(p []byte) (n int, err error) {
+	return (*zlib.Writer)(zw).Write(p)
+}
+func (zw *ZlibWriter) Flush() error {
+	return (*zlib.Writer)(zw).Flush()
+}
+func (zw *ZlibWriter) Close() error {
+	return (*zlib.Writer)(zw).Close()
+}
+
+func NewZlibWriterLevel(level int) *ZlibWriter {
+	zw, _ := zlib.NewWriterLevel(nil, level)
+	return (*ZlibWriter)(zw)
+}
 
 const MaxDataLength = 2097152
 
@@ -41,9 +61,9 @@ func (p Packet) Scan(fields ...FieldDecoder) error {
 }
 
 // Pack 打包一个数据包
-func (p *Packet) Pack(w io.Writer, threshold int) error {
+func (p *Packet) Pack(w io.Writer, threshold int, zw *ZlibWriter) error {
 	if threshold >= 0 {
-		return p.packWithCompression(w, threshold)
+		return p.packWithCompression(w, threshold, zw)
 	} else {
 		return p.packWithoutCompression(w)
 	}
@@ -75,7 +95,7 @@ func (p *Packet) packWithoutCompression(w io.Writer) error {
 	return nil
 }
 
-func (p *Packet) packWithCompression(w io.Writer, threshold int) error {
+func (p *Packet) packWithCompression(w io.Writer, threshold int, zw *ZlibWriter) error {
 	buff := bufPool.Get().(*bytes.Buffer)
 	defer bufPool.Put(buff)
 	buff.Reset()
@@ -105,8 +125,9 @@ func (p *Packet) packWithCompression(w io.Writer, threshold int) error {
 		}
 	} else {
 		// zw := zlib.NewWriter(buff)
-		zw := zlibWriterPool.Get(buff)
-		defer zlibWriterPool.Return(zw)
+		// zw := zlibWriterPool.Get(buff)
+		// defer zlibWriterPool.Return(zw)
+		zw.Reset(buff)
 		n1, err := VarInt(p.ID).WriteTo(zw)
 		if err != nil {
 			return err