--- a/net/packet/packet.go
+++ b/net/packet/packet.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"fmt"
 	"io"
+	"sync"
 )
 
 const MaxDataLength = 2097152
@@ -25,8 +26,7 @@ func Marshal(id int32, fields ...FieldEncoder) (pk Packet) {
 
 //Scan decode the packet and fill data into fields
 func (p Packet) Scan(fields ...FieldDecoder) error {
-	r := buffReaderPool.Get(p.Data)
-	defer buffReaderPool.Return(r)
+	r := bytes.NewReader(p.Data)
 	for _, v := range fields {
 		_, err := v.ReadFrom(r)
 		if err != nil {
@@ -34,6 +34,12 @@ func (p Packet) Scan(fields ...FieldDecoder) error {
 		}
 	}
 	return nil
+}
+
+var bufPool = sync.Pool{
+	New: func() interface{} {
+		return new(bytes.Buffer)
+	},
 }
 
 // Pack 打包一个数据包
@@ -196,10 +202,7 @@ func (p *Packet) unpackWithCompression(r io.Reader, threshold int) error {
 	if err != nil {
 		return err
 	}
-	// r = bytes.NewReader(buff.Bytes())
-	br := buffReaderPool.Get(buff.Bytes())
-	r = br
-	defer buffReaderPool.Return(br)
+	r = bytes.NewReader(buff.Bytes())
 
 	var DataLength VarInt
 	n2, err := DataLength.ReadFrom(r)