As mentioned in #209,
I add `KeyManager` for caching RSA keypair,
not yet add the option for `prevent-proxy-connections: true` right now.

And this PR also include some performance enhancement in packet level, which:
1. use `sync.Pool` for `bytes.NewReader()` in `Packet.Scan()`
2. switch standard library `compress/zlib` to third party `github.com/klauspost/compress/zlib`
3. add option for setting compress level