@Tnze 2022-07-20
I don't believe the first change can improve the performance.

`bytes.NewReader` doesn't allocate any memory on heap. Using `sync.Pool` in this case only cause the variables escape from stack(or registers) to heap, which decreases the performance actually.