@cs8425 2022-07-21
> I don't believe the first change can improve the performance.
> 
> `bytes.NewReader` doesn't allocate any memory on heap. Using `sync.Pool` in this case only cause the variables escape from stack(or registers) to heap, which decreases the performance actually.

Actually only save 1 allocate,
and seems performance nearly same...

I add this code in `net/packet/packet_test.go`
```golang
func BenchmarkPacketScan(b *testing.B) {
	p := pk.Packet{ID: 0x24, Data: testJoinGameData}
	var (
		EID            pk.Int
		Hardcore       pk.Boolean
		Gamemode       pk.UnsignedByte
		PreGamemode    pk.Byte
		WorldNames     = make([]pk.Identifier, 0) // This cannot replace with "var DimensionNames []pk.Identifier" because "nil" has no type information
		DimensionCodec struct {
			DimensionType interface{} `nbt:"minecraft:dimension_type"`
			WorldgenBiome interface{} `nbt:"minecraft:worldgen/biome"`
		}
		Dimension                 interface{}
		WorldName                 pk.Identifier
		HashedSeed                pk.Long
		MaxPlayers                pk.VarInt
		ViewDistance              pk.VarInt
		RDI, ERS, IsDebug, IsFlat pk.Boolean
	)

	b.SetBytes(int64(len(p.Data)))
	b.ResetTimer()
	b.ReportAllocs()

	for i := 0; i < b.N; i++ {
		err := p.Scan(
			&EID,
			&Hardcore,
			&Gamemode,
			&PreGamemode,
			pk.Array(&WorldNames),
			pk.NBT(&DimensionCodec),
			pk.NBT(&Dimension),
			&WorldName,
			&HashedSeed,
			&MaxPlayers,
			&ViewDistance,
			&RDI, &ERS, &IsDebug, &IsFlat,
		)
		if err != nil {
			b.Fatalf("err = %d; want <nil>", err)
		}
	}
}
```

benchmark result:

without pool:
```
$ go test -bench=. -benchmem -cpu=1,2,4,8 -run=^#
goos: linux
goarch: amd64
pkg: github.com/Tnze/go-mc/net/packet
cpu: Intel(R) Core(TM) i3-10100 CPU @ 3.60GHz
BenchmarkPacketScan     	    2385	    489661 ns/op	  63.00 MB/s	  224318 B/op	   12713 allocs/op
BenchmarkPacketScan-2   	    2406	    487766 ns/op	  63.24 MB/s	  224336 B/op	   12713 allocs/op
BenchmarkPacketScan-4   	    2433	    496988 ns/op	  62.07 MB/s	  224325 B/op	   12713 allocs/op
BenchmarkPacketScan-8   	    2402	    493389 ns/op	  62.52 MB/s	  224336 B/op	   12713 allocs/op
PASS
ok  	github.com/Tnze/go-mc/net/packet	4.945s
```
```
$ go test -bench=. -benchmem -cpu=1,2,4,8,16 -run=^#
goos: linux
goarch: amd64
pkg: github.com/Tnze/go-mc/net/packet
cpu: AMD Ryzen 7 5700G with Radeon Graphics         
BenchmarkPacketScan       	    2367	    465953 ns/op	  66.20 MB/s	  224331 B/op	   12713 allocs/op
BenchmarkPacketScan-2     	    2185	    486443 ns/op	  63.41 MB/s	  224340 B/op	   12713 allocs/op
BenchmarkPacketScan-4     	    2139	    498789 ns/op	  61.84 MB/s	  224335 B/op	   12713 allocs/op
BenchmarkPacketScan-8     	    2175	    524318 ns/op	  58.83 MB/s	  224333 B/op	   12713 allocs/op
BenchmarkPacketScan-16    	    2128	    529007 ns/op	  58.31 MB/s	  224338 B/op	   12713 allocs/op
PASS
ok  	github.com/Tnze/go-mc/net/packet	5.789s
```

with pool:
```
$ go test -bench=. -benchmem -cpu=1,2,4,8 -run=^#
goos: linux
goarch: amd64
pkg: github.com/Tnze/go-mc/net/packet
cpu: Intel(R) Core(TM) i3-10100 CPU @ 3.60GHz
BenchmarkPacketScan     	    2392	    491476 ns/op	  62.76 MB/s	  224283 B/op	   12712 allocs/op
BenchmarkPacketScan-2   	    2421	    485727 ns/op	  63.51 MB/s	  224301 B/op	   12712 allocs/op
BenchmarkPacketScan-4   	    2383	    493056 ns/op	  62.56 MB/s	  224308 B/op	   12712 allocs/op
BenchmarkPacketScan-8   	    2386	    505003 ns/op	  61.08 MB/s	  224354 B/op	   12712 allocs/op
PASS
ok  	github.com/Tnze/go-mc/net/packet	4.942s
```
```
$ go test -bench=. -benchmem -cpu=1,2,4,8,16 -run=^#
goos: linux
goarch: amd64
pkg: github.com/Tnze/go-mc/net/packet
cpu: AMD Ryzen 7 5700G with Radeon Graphics         
BenchmarkPacketScan       	    2390	    456629 ns/op	  67.55 MB/s	  224288 B/op	   12712 allocs/op
BenchmarkPacketScan-2     	    2342	    482222 ns/op	  63.97 MB/s	  224307 B/op	   12712 allocs/op
BenchmarkPacketScan-4     	    2216	    522766 ns/op	  59.01 MB/s	  224313 B/op	   12712 allocs/op
BenchmarkPacketScan-8     	    2210	    515063 ns/op	  59.89 MB/s	  224344 B/op	   12712 allocs/op
BenchmarkPacketScan-16    	    2276	    523909 ns/op	  58.88 MB/s	  224412 B/op	   12712 allocs/op
PASS
ok  	github.com/Tnze/go-mc/net/packet	5.986s
```