@Tnze 2022-07-21
> and seems performance nearly same...

Remove the pool.