@Tnze 2022-07-21
Let's do what we talk about in #209 first.

(By simply store the server's key in the LoginHandler. There is no need to use interface if we have only one implementation.

I kept my opinion of switching the `zlib` lib, and you can open a new issue for that if you like.